Move rule-specific schema tests
